Split into multiple verticles
